Replaced default `None` with `_UNSET`
Considering `None` can be default value for argument, `_UNSET` was
created to handle the case when argument IS NOT SET. This especially is
needed where we have function wrappers on top of "shadow" functions.
See `integrify.kapital.client.KapitalClientClass.pay_with_saved_card`
which uses 3 "shadow" functions, and passed `None` as a argument. If we
had pydantic default value for that field, that would not have
triggered. Now, we just filter out all _UNSET args before "sending" them
to pydantic model, thus triggering defaults if needed.